Add next page link for second last page.  fixes https://github.com/onaio/fhir-gateway-extension/issues/45

+++ plugins/src/main/java/org/smartregister/fhir/gateway/plugins/SyncAccessDecision.java
@@ -373,7 +373,7 @@ public class SyncAccessDecision implements AccessDecision {
         resultBundle.setTotal(totalEntries);
 
         // add pagination links
-        int nextPage = page < totalEntries / count ? page + 1 : 0; // 0 indicates no next page
+        int nextPage = page < ((float)totalEntries / count) ? page + 1 : 0; // 0 indicates no next page
         int prevPage = page > 1 ? page - 1 : 0; // 0 indicates no previous page
 
         Bundle.BundleLinkComponent selfLink = new Bundle.BundleLinkComponent();
